outdated/requirements.txt
numpy==2.0.0
torch==2.3.1
pandas==2.2.2
tqdm==4.66.4
python-dotenv==1.0.1
pydantic==2.7.4
wandb==0.17.2
lightning==2.3.0
torchvision==0.18.1
PyYAML==6.0.1